Move resource limits checking to a separate package
